@@ -1,5 +1,5 @@
 plugins {
-    id 'org.jetbrains.kotlin.jvm' version '1.3.11'
+    id 'org.jetbrains.kotlin.jvm' version '1.6.0'
     id 'application'
 }
 
@@ -36,7 +36,7 @@ compileKotlin {
     targetCompatibility = JavaVersion.VERSION_11
 
     kotlinOptions {
-        jvmTarget = "1.8"
+        jvmTarget = "11"
     }
 }
 
@@ -45,7 +45,7 @@ compileTestKotlin {
     targetCompatibility = JavaVersion.VERSION_11
 
     kotlinOptions {
-        jvmTarget = "1.8"
+        jvmTarget = "11"
     }
 }
 
