@@ -10,8 +10,8 @@ fun <E> MutableList<E>.addUnique(element: E)
 
 fun <K: Comparable<K>, V> Map<K, V>.getSortedValues(): List<V> = entries.sortedBy { it.key }.map { it.value }
 
-fun List<Int>.minOrZero() = min() ?: 0
-fun List<Int>.maxOrZero() = max() ?: 0
+fun List<Int>.minOrZero() = minOrNull() ?: 0
+fun List<Int>.maxOrZero() = maxOrNull() ?: 0
 
 inline fun <T, R : Comparable<R>> Iterable<T>.sortedBy(descending: Boolean, crossinline selector: (T) -> R?): List<T> {
     return if (descending) this.sortedByDescending(selector) else this.sortedBy(selector)
